shark-detector/model.py
import torch
from PIL import Image
import torchvision.transforms as transforms

# Load your model here (example)
# model = torch.load('your_model.pth', map_location=torch.device('cpu'))
# model.eval()

def predict_shark(image_path):
    """
    Dummy function to simulate shark detection.
    Replace with your actual model inference code.
    """
    try:
        # Example image preprocessing
        image = Image.open(image_path).convert('RGB')
        transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
        ])
        input_tensor = transform(image).unsqueeze(0)

        # Example prediction (replace with: outputs = model(input_tensor))
        # For demonstration, we just return a dummy result
        result = {"detected": True, "species": "Great White Shark", "confidence": 0.95}
        return result
    except Exception as e:
        return {"error": str(e)}
